defining variable

+++ 0x00-python_variable_annotations/4-define_variables.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """Type annotation."""
 a: int = 1
-pi: int = 3.14
+pi: float = 3.14
 i_understand_annotations: bool = True
 school: str = "Holberton"
